refactor: move `luaLSType`
